Give a human friendly exception message when we encounter an issue here
An example of when this exception can be triggered is when installing nokogiri under the generic ruby platform. It has a dependency of "mini_portile2", but when installing the gem, it will actually install a platform specific version of nokogiri if one exists (for instance arm64-darwin) which do not have "mini_portile2" as a dependency. Therefore the Gemfile.lock will list "mini_portile2" in specs:, but it won't actually be installed on your system. Therefore "bundle list --paths" for that gem will not exist.